maxPostCount & maxCommentCount
